+++ src/main/kotlin/com/github/blarc/sops/intellij/plugin/settings/AppSettingsConfigurable.kt
@@ -3,6 +3,7 @@ package com.github.blarc.sops.intellij.plugin.settings
 import com.github.blarc.sops.intellij.plugin.SopsBundle.message
 import com.intellij.execution.configuration.EnvironmentVariablesComponent
 import com.intellij.openapi.options.BoundConfigurable
+import com.intellij.ui.dsl.builder.Align
 import com.intellij.ui.dsl.builder.bindText
 import com.intellij.ui.dsl.builder.panel
 import com.intellij.ui.dsl.builder.toMutableProperty
@@ -13,15 +14,15 @@ class AppSettingsConfigurable : BoundConfigurable(message("name")) {
             label(message("settings.path"))
                 .widthGroup("label")
             textField()
-                .widthGroup("input")
+                .align(Align.FILL)
                 .bindText(AppSettings.instance::sopsPath)
         }
 
         row {
-            label(message("settings.path"))
+            label(message("settings.environment"))
                 .widthGroup("label")
             cell(EnvironmentVariablesComponent())
-                .widthGroup("input")
+                .align(Align.FILL)
                 .applyToComponent {
                     text = ""
                 }